Use size_t for the number of shared memory bytes.

+++ hoomd/hpmc/IntegratorHPMCMonoGPUDepletants.cuh
@@ -702,14 +702,12 @@ void depletants_launcher(const hpmc_args_t& args,
         unsigned int max_queue_size = n_groups * tpp;
         unsigned int max_depletant_queue_size = n_groups;
 
-        const unsigned int min_shared_bytes
-            = static_cast<unsigned int>(args.num_types * sizeof(typename Shape::param_type)
-                                        + args.overlap_idx.getNumElements() * sizeof(unsigned int));
+        const size_t min_shared_bytes = args.num_types * sizeof(typename Shape::param_type)
+                                        + args.overlap_idx.getNumElements() * sizeof(unsigned int);
 
-        unsigned int shared_bytes = static_cast<unsigned int>(
-            n_groups * (sizeof(Scalar4) + sizeof(Scalar3))
-            + max_queue_size * 2 * sizeof(unsigned int)
-            + max_depletant_queue_size * sizeof(unsigned int) + min_shared_bytes);
+        size_t shared_bytes = n_groups * (sizeof(Scalar4) + sizeof(Scalar3))
+                              + max_queue_size * 2 * sizeof(unsigned int)
+                              + max_depletant_queue_size * sizeof(unsigned int) + min_shared_bytes;
 
         if (min_shared_bytes >= args.devprop.sharedMemPerBlock)
             throw std::runtime_error("Insufficient shared memory for HPMC kernel: reduce number of "
@@ -729,10 +727,9 @@ void depletants_launcher(const hpmc_args_t& args,
             max_queue_size = n_groups * tpp;
             max_depletant_queue_size = n_groups;
 
-            shared_bytes = static_cast<unsigned int>(
-                n_groups * (sizeof(Scalar4) + sizeof(Scalar3))
-                + max_queue_size * 2 * sizeof(unsigned int)
-                + max_depletant_queue_size * sizeof(unsigned int) + min_shared_bytes);
+            shared_bytes = n_groups * (sizeof(Scalar4) + sizeof(Scalar3))
+                           + max_queue_size * 2 * sizeof(unsigned int)
+                           + max_depletant_queue_size * sizeof(unsigned int) + min_shared_bytes;
             }
 
         // determine dynamically requested shared memory
